hicomp.py - comparison histograms for processed data
Initial version of this function. Generates histograms of the raw counts
compared to the counts in a comparison file (generally the bias and flat
field corrected version). Each FITS extension from a given file is
plotted with the same color (and some transparancy) to provide a quick
assessment of the qualitative differences in the counts between two
images.
Plots are saved as png files.

Currently this is hard-coded for the Bok 90Prime's 16 extension FITS
files, but will be extended in the future to arbitrary numbers of FITS
extensions.